requirements/base.txt
Django==3.0.8
django-environ==0.4.5
psycopg2==2.8.5
django-choices==1.7.1
django-crispy-forms==1.9.2
django-compressor==2.4
django-dramatiq==0.9.1
redis==3.5.3
feedparser==5.2.1
whitenoise==5.1.0
pika==1.1.0
sphinx==3.1.2
sphinx-autobuild==0.7.1